chore: move the data fetching to a hook
